♻️ refactor: AuctionController product_id -> productId로 리팩토링
