+++ src/main/java/com/sparta/itsmine/domain/qna/controller/QnaController.java
@@ -9,10 +9,10 @@ import static com.sparta.itsmine.global.common.ResponseCodeEnum.SUCCESS_UPDATE_Q
 import com.sparta.itsmine.domain.qna.dto.GetQnaResponseDto;
 import com.sparta.itsmine.domain.qna.dto.QnaRequestDto;
 import com.sparta.itsmine.domain.qna.service.QnaService;
-import com.sparta.itsmine.domain.security.UserDetailsImpl;
 import com.sparta.itsmine.domain.user.entity.User;
 import com.sparta.itsmine.global.common.HttpResponseDto;
 import com.sparta.itsmine.global.common.ResponseUtils;
+import com.sparta.itsmine.global.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
